+++ src/collective/lazysizes/transform.py
@@ -3,8 +3,10 @@ from collective.lazysizes.interfaces import ILazySizesSettings
 from collective.lazysizes.logger import logger
 from lxml import etree
 from plone import api
+from plone.registry.interfaces import IRegistry
 from plone.transformchain.interfaces import ITransform
 from repoze.xmliter.utils import getHTMLSerializer
+from zope.component import getUtility
 from zope.interface import implementer
 
 
@@ -154,16 +156,20 @@ class LazySizesTransform(object):
         return
 
     def transformIterable(self, result, encoding):
+        registry = getUtility(IRegistry)
+        settings = registry.forInterface(ILazySizesSettings)
+
+        # we apply the transform always for anonymous users
         if not api.user.is_anonymous():
-            return
+            # the user is authenticated, check if enabled
+            if not settings.lazyload_authenticated:
+                return  # transform not enabled for authenticated users
 
         result = self._parse(result)
         if result is None:
             return
 
-        record = ILazySizesSettings.__identifier__ + '.css_class_blacklist'
-        blacklist = api.portal.get_registry_record(record)
-        blacklist = self._blacklist(result, blacklist)
+        blacklist = self._blacklist(result, settings.css_class_blacklist)
 
         path = '{0}//img|{0}//iframe|{0}//blockquote'.format(ROOT_SELECTOR)
         for el in result.tree.xpath(path):